Refactor refreshrate handling from int to float.
With SDL 2 the refreshrate handling was rather limited by SDL 2 only
supporting integer refreshrates, making it impossible to represent the
common 59.95hz and other non-integer refreshrates. Quake II works around
this by assuming a refreshrate of 2hz higher than returned by SDL.

SDL 3 finally supports float refreshrates. Refactor the internal API to
also use floats.

This should be a no-op since the 'refreshrate * 1.02' logic stayes in
place for the time being. This can be reevaluated at a later time for
the SDL 3 build, but needs testing.

As a nice side effect fix the `vid_displayrefreshrate` cvar. It was
broken in both SDL 2 and 3 since it was always overwritten with the
actual display refresh rate.

+++ src/client/vid/header/vid.h
@@ -54,7 +54,7 @@ void	VID_MenuDraw(void);
 const char *VID_MenuKey(int);
 
 // Stuff provided by platform backend.
-extern int glimp_refreshRate;
+extern float glimp_refreshRate;
 
 const char **GLimp_GetDisplayIndices(void);
 int GLimp_GetWindowDisplayIndex(void);
@@ -64,7 +64,7 @@ void GLimp_Shutdown(void);
 qboolean GLimp_InitGraphics(int fullscreen, int *pwidth, int *pheight);
 void GLimp_ShutdownGraphics(void);
 void GLimp_GrabInput(qboolean grab);
-int GLimp_GetRefreshRate(void);
+float GLimp_GetRefreshRate(void);
 qboolean GLimp_GetDesktopMode(int *pwidth, int *pheight);
 
 #endif
